pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "strategic-consultant-agent"
version = "0.1.0"
description = "Multi-agent strategic decision support system using Google ADK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.19.0",
    "google-genai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-cov>=7.0.0",
    "pytest-asyncio>=1.0.0",
    "black>=25.11.0",
    "pylint>=4.0.0",
    "mypy>=1.14.0",
]

[tool.setuptools]
packages = ["strategic_consultant_agent"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.black]
line-length = 88
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
